done adding

+++ products/products/src/main/java/com/arsh/products/service/ProductService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -24,11 +25,19 @@ public class ProductService {
 
 
     public Product getProductById(Long id) {
-        Product dummyProduct = new Product();
-        dummyProduct.setId(0L);
-        dummyProduct.setTitle("product not found");
-        dummyProduct.setPrice(0);
-        return productRepository.findById(id).orElse(dummyProduct);
+//        Product dummyProduct = new Product();
+//        dummyProduct.setId(0L);
+//        dummyProduct.setTitle("product not found");
+//        dummyProduct.setPrice(0);
+
+//        Optional<Product> optionalProduct = productRepository.findById(id);
+//        Product product;
+//        if (optionalProduct.isPresent()) {
+//            product = optionalProduct.get();
+//        }else {
+//            product = null;
+//        }
+        return productRepository.findById(id).orElse(null);
     }
 
     public List<Product> getProductsByTitle(String searchTerm) {
@@ -41,13 +50,13 @@ public class ProductService {
 
 
     public List<Product> getProductsByTitleAndPriceRange(String searchTerm, int startPrice, int endPrice) {
-        return productRepository.findByTitleContainingIgnoreCaseAndPriceBetween(searchTerm,startPrice,endPrice);
+        return productRepository.findByTitleContainingIgnoreCaseAndPriceBetween(searchTerm, startPrice, endPrice);
 
     }
 
 
     public List<Product> getProductsByBrand(String brand) {
-    return productRepository.findByBrandIgnoreCase(brand);
+        return productRepository.findByBrandIgnoreCase(brand);
     }
 
     public List<Product> getProductsByCategory(String category) {
@@ -56,11 +65,11 @@ public class ProductService {
 
 
     public List<Product> getProductsByBrandCategory(String brand, String category) {
-    return productRepository.findByBrandAndCategoryIgnoreCase(brand,category);
+        return productRepository.findByBrandAndCategoryIgnoreCase(brand, category);
     }
 
     public List<Product> getProductsByTitleBrandCategory(String searchTerm, String brand, String category) {
-    return productRepository.findByTitleAndBrandAndCategoryIgnoreCase(searchTerm,brand,category);
+        return productRepository.findByTitleAndBrandAndCategoryIgnoreCase(searchTerm, brand, category);
     }
 
     public List<Product> getProductsByCategory0rderByTitle(String category) {
@@ -68,10 +77,11 @@ public class ProductService {
     }
 
     public List<Product> getProductsByPriceDesc() {
-     return productRepository.findAllProductsByOrderByPriceDesc();
+        return productRepository.findAllProductsByOrderByPriceDesc();
     }
-    public List<Product> findByBrandAndPriceBetween(String brand,double mintPrice, double maxPrice) {
-        return productRepository.findByBrandAndPriceBetween(brand,mintPrice,maxPrice);
+
+    public List<Product> findByBrandAndPriceBetween(String brand, double mintPrice, double maxPrice) {
+        return productRepository.findByBrandAndPriceBetween(brand, mintPrice, maxPrice);
     }
 
 
@@ -80,23 +90,33 @@ public class ProductService {
     }
 
     public Product createProduct(Product product) {
-        return productRepository.save(product);
+        if (product == null) {
+            return null;
+        } else {
+            return productRepository.save(product);
+        }
     }
 
     public Product updateProduct(Long id, Product productDetails) {
-
-        Product product = productRepository.findById(id).orElseThrow();
-        product.setTitle(productDetails.getTitle());
-        product.setPrice(productDetails.getPrice());
-        return productRepository.save(product);
-
+        Optional<Product> productOptional = productRepository.findById(id);
+        if (productOptional.isEmpty()) {
+            return null;
+        } else {
+            Product product = productOptional.get();
+            product.setTitle(productDetails.getTitle());
+            product.setPrice(productDetails.getPrice());
+            product.setCategory(productDetails.getCategory());
+            product.setBrand(productDetails.getBrand());
+            return productRepository.save(product);
+        }
     }
 
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+    public boolean deleteProduct(Long id) {
+        if (!productRepository.existsById(id)) {
+            return false;
+        } else {
+            productRepository.deleteById(id);
+            return true;
+        }
     }
-
-
 }
-
-
